file-storage: extract atomic write helper into its own module
Pure refactor: write_file_atomically/temp_path_for/backup_path_for
move from migration.rs into a new atomic_write module so the normal
data write path can reuse them in the next commit. No behavior
change.
