+++ storages/file-storage/src/migration.rs
@@ -1,7 +1,7 @@
 mod v1_to_v2;
 
 use {
-    crate::{FileStorage, ResultExt},
+    crate::{FileStorage, ResultExt, atomic_write::write_file_atomically},
     gluesql_core::{
         data::Schema,
         error::{Error, Result},
@@ -10,10 +10,8 @@ use {
         convert::AsRef,
         ffi::OsStr,
         fs,
-        io::Write,
         path::{Path, PathBuf},
     },
-    uuid::Uuid,
 };
 
 pub const FILE_STORAGE_FORMAT_VERSION: u32 = 2;
@@ -140,58 +138,6 @@ fn list_schema_paths(path: &Path) -> Result<Vec<PathBuf>> {
 fn read_schema_file(path: &Path) -> Result<SchemaFile> {
     let data = fs::read_to_string(path).map_storage_err()?;
     parse_schema_file(data)
-}
-
-fn write_file_atomically(path: &Path, data: &str) -> Result<()> {
-    let temp_path = temp_path_for(path);
-    let backup_path = backup_path_for(path);
-    let has_existing_target = path.exists();
-
-    let mut file = fs::File::create(&temp_path).map_storage_err()?;
-    file.write_all(data.as_bytes()).map_storage_err()?;
-    file.sync_all().map_storage_err()?;
-    drop(file);
-
-    if has_existing_target && let Err(backup_err) = fs::rename(path, &backup_path).map_storage_err()
-    {
-        let _ = fs::remove_file(&temp_path);
-        return Err(backup_err);
-    }
-
-    if let Err(target_rename_err) = fs::rename(&temp_path, path).map_storage_err() {
-        let _ = fs::remove_file(&temp_path);
-        if has_existing_target
-            && let Err(restore_err) = fs::rename(&backup_path, path).map_storage_err()
-        {
-            return Err(Error::StorageMsg(format!(
-                "[FileStorage] failed to atomically replace '{}': {target_rename_err}; and failed to restore backup '{}': {restore_err}",
-                path.display(),
-                backup_path.display()
-            )));
-        }
-
-        return Err(target_rename_err);
-    }
-
-    if has_existing_target {
-        let _ = fs::remove_file(&backup_path);
-    }
-
-    Ok(())
-}
-
-fn temp_path_for(path: &Path) -> PathBuf {
-    let extension = path.extension().and_then(OsStr::to_str).unwrap_or_default();
-    let suffix = Uuid::now_v7();
-
-    path.with_extension(format!("{extension}.tmp-{suffix}"))
-}
-
-fn backup_path_for(path: &Path) -> PathBuf {
-    let extension = path.extension().and_then(OsStr::to_str).unwrap_or_default();
-    let suffix = Uuid::now_v7();
-
-    path.with_extension(format!("{extension}.bak-{suffix}"))
 }
 
 fn parse_schema_file(data: String) -> Result<SchemaFile> {
